pyproject.toml
[tool.poetry]
name = "sb1-noticeboard"
version = "0.1.0"
description = ""
authors = ["SChuprinenko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
